% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_branch_table.R
\name{branches}
\alias{branches}
\title{Build branch table}
\usage{
branches(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{overwrite}{Logical. Rebuild an existing branch table?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Builds the internal branch representation from a prepared phylogenetic tree.
}
