% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_geometry_index.R
\name{build_geometry_index}
\alias{build_geometry_index}
\title{Build geometry index}
\usage{
build_geometry_index(pb, id_column, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A pb_project object.}

\item{id_column}{Name of the unique feature identifier.}

\item{overwrite}{Rebuild an existing cache.}

\item{verbose}{Print progress messages.}
}
\value{
Updated pb_project.
}
\description{
Build an internal lookup index for basin geometries.
}
