% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_from_geometry_engine.R
\name{community_from_geometry_engine}
\alias{community_from_geometry_engine}
\title{Build a community matrix from geometry attributes}
\usage{
community_from_geometry_engine(data, site_id, species_columns)
}
\arguments{
\item{data}{A data.frame containing geometry attributes.}

\item{site_id}{Name of the site identifier column.}

\item{species_columns}{Character vector of species columns.}
}
\value{
A numeric presence/absence matrix with site identifiers as row names.
}
\description{
Internal engine used by build_community_from_geometry().
}
\keyword{internal}
