% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_descendants.R
\name{compute_branch_descendants}
\alias{compute_branch_descendants}
\title{Compute branch descendants}
\usage{
compute_branch_descendants(pb, store_descendants = FALSE)
}
\arguments{
\item{pb}{A \code{pb_project} object.}

\item{store_descendants}{Logical; should complete descendant-tip lists be
stored in \code{pb$branches$cache$descendants}?}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes the number of descendant tips subtended by every branch in the tree.
Optionally stores the complete descendant tip lists in the branch cache.
}
