% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_ranges.R
\name{compute_branch_ranges}
\alias{compute_branch_ranges}
\title{Compute branch ranges}
\usage{
compute_branch_ranges(pb)
}
\arguments{
\item{pb}{A PhyloBasins project.}
}
\value{
Updated project.
}
\description{
Computes the number of sites in which every branch occurs.
}
