% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pd.R
\name{compute_pd}
\alias{compute_pd}
\title{Compute Faith's Phylogenetic Diversity}
\usage{
compute_pd(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{overwrite}{Logical. Recompute existing values?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes Faith's phylogenetic diversity (PD) for all communities.
}
