% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pe.R
\name{compute_pe}
\alias{compute_pe}
\title{Compute phylogenetic endemism}
\usage{
compute_pe(pb)
}
\arguments{
\item{pb}{A prepared PhyloBasins project.}
}
\value{
Updated project.
}
\description{
Computes phylogenetic endemism (PE) for every site.
}
