% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_richness.R
\name{compute_richness}
\alias{compute_richness}
\title{Compute species richness}
\usage{
compute_richness(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A PhyloBasins project.}

\item{overwrite}{Logical. Recompute richness if it already exists.}

\item{verbose}{Logical. Print progress messages.}
}
\value{
Updated PhyloBasins project.
}
\description{
Computes species richness (number of taxa present) for every community.
}
