% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rpe.R
\name{compute_rpe}
\alias{compute_rpe}
\title{Compute Relative Phylogenetic Endemism}
\usage{
compute_rpe(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{overwrite}{Logical. Recompute existing values?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes Relative Phylogenetic Endemism (RPE) as the ratio between
observed PE and PE calculated on an equal-branch-length tree having the
same total tree length.
}
