% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_subtree_lengths.R
\name{compute_subtree_lengths}
\alias{compute_subtree_lengths}
\title{Compute subtree lengths}
\usage{
compute_subtree_lengths(pb)
}
\arguments{
\item{pb}{A prepared \code{pb_project}.}
}
\value{
Updated \code{pb_project}.
}
\description{
Calculates the cumulative branch length for every branch in the tree.
Each subtree length equals the length of the focal branch plus the
lengths of all descendant branches.
}
