% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_turnover.R
\name{compute_turnover}
\alias{compute_turnover}
\title{Compute community turnover}
\usage{
compute_turnover(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A PhyloBasins project.}

\item{overwrite}{Logical. Recompute existing values?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes the mean Jaccard dissimilarity (community turnover) for every site.
Turnover is calculated as the mean binary Jaccard distance between each site
and all other sites.
}
