% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_species_columns.R
\name{detect_species_columns}
\alias{detect_species_columns}
\title{Detect species columns}
\usage{
detect_species_columns(
  data,
  species_columns = NULL,
  first_species = NULL,
  last_species = NULL
)
}
\arguments{
\item{data}{A data.frame.}

\item{species_columns}{Optional character vector of species columns.}

\item{first_species}{Optional name of the first species column.}

\item{last_species}{Optional name of the last species column.}
}
\value{
Character vector of species column names.
}
\description{
Determines which columns of a geometry attribute table represent species.
}
\details{
Species columns may be supplied explicitly or defined by the first and
last species column names.
}
\keyword{internal}
