% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_tree}
\alias{example_tree}
\title{Example phylogenetic tree}
\format{
An object of class \code{"phylo"}.
}
\source{
Generated for package examples.
}
\usage{
example_tree
}
\description{
A small rooted phylogenetic tree containing five species.
}
\keyword{datasets}
