% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side.}

\item{y}{Right-hand side.}
}
\value{
Either \code{x} or \code{y}.
}
\description{
Returns the left-hand side unless it is NULL, otherwise returns the
right-hand side.
}
\examples{
NULL \%||\% 1
5 \%||\% 1
}
\keyword{internal}
