% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{is.pb_community}
\alias{is.pb_community}
\title{Test whether an object is a pb_community}
\usage{
is.pb_community(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical.
}
\description{
Test whether an object is a pb_community
}
