% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_metric_to_geometry.R
\name{join_metric_to_geometry}
\alias{join_metric_to_geometry}
\title{Join metrics to geometry}
\usage{
join_metric_to_geometry(
  pb,
  metrics = NULL,
  copy_geometry = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{pb}{A pb_project object.}

\item{metrics}{Character vector of metric names.
If NULL, all available metrics are joined.}

\item{copy_geometry}{Logical.
If TRUE (default), return a copy of the geometry.
If FALSE, modify the geometry stored in the project.}

\item{overwrite}{Replace existing attribute columns.}
}
\value{
If \code{copy_geometry = TRUE}, an \strong{sf} object.
Otherwise an updated \strong{pb_project}.
}
\description{
Attach one or more computed metrics to the geometry stored in a
PhyloBasins project.
}
