% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_ranges.R
\name{new_branch_ranges}
\alias{new_branch_ranges}
\title{Create an empty branch ranges object}
\usage{
new_branch_ranges()
}
\value{
A new \code{pb_branch_ranges} object.
}
\description{
Create an empty branch ranges object
}
\keyword{internal}
