% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{new_geometry}
\alias{new_geometry}
\title{Create a geometry object}
\usage{
new_geometry(
  sf = NULL,
  file = NA_character_,
  loaded = FALSE,
  validation = list(valid = FALSE),
  metadata = list(),
  cache = list(id_index = NULL, centroids = NULL, neighbours = NULL, spatial_index =
    NULL)
)
}
\arguments{
\item{sf}{An sf object or NULL.}

\item{file}{Source filename.}

\item{loaded}{Logical.}

\item{validation}{Validation information.}

\item{metadata}{Metadata.}

\item{cache}{Internal cache.}
}
\value{
A pb_geometry object.
}
\description{
Internal constructor for spatial geometry.
}
\keyword{internal}
