% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalescing}
\alias{null-coalescing}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side.}

\item{y}{Right-hand side.}
}
\value{
Either \code{x} or \code{y}.
}
\description{
Returns \code{x} unless it is \code{NULL}, otherwise returns \code{y}.
}
\keyword{internal}
