% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_project.R
\name{pb_project}
\alias{pb_project}
\title{Create a new PhyloBasins project}
\usage{
pb_project()
}
\value{
An object of class \code{"pb_project"}.
}
\description{
Creates an empty PhyloBasins project object.
}
\details{
The returned object stores all data, metadata and intermediate
results generated during a PhyloBasins workflow.
}
\examples{
pb <- pb_project()

}
