% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-pb_test_project.R
\name{pb_test_project}
\alias{pb_test_project}
\title{Create a test project}
\usage{
pb_test_project(
  stage = c("empty", "tree", "prepared_tree", "branches", "community",
    "site_branch_matrix", "branch_ranges", "pd", "pe", "rpe")
)
}
\arguments{
\item{stage}{Character. One of:

"empty"
"tree"
"prepared_tree"
"branches"
"community"
"site_branch_matrix"
"branch_ranges"
"pd"
"pe"
"rpe"}
}
\value{
A valid pb_project object.
}
\description{
Creates a minimal pb_project object and optionally executes the pipeline
up to a selected stage.
}
\keyword{internal}
