% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geometry.R
\name{plot_geometry}
\alias{plot_geometry}
\title{Plot geometry}
\usage{
plot_geometry(
  pb,
  fill = NULL,
  border = "grey60",
  palette = "viridis",
  legend_title = ggplot2::waiver(),
  na_colour = "grey90",
  linewidth = 0.15
)
}
\arguments{
\item{pb}{A pb_project object.}

\item{fill}{Optional numeric vector used for colouring.}

\item{border}{Border colour.}

\item{palette}{Colour palette.}

\item{legend_title}{Legend title.}

\item{na_colour}{Colour for missing values.}

\item{linewidth}{Polygon border width.}
}
\value{
A ggplot object.
}
\description{
Plot the geometry stored in a PhyloBasins project.
}
