% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pe.R
\name{plot_pe}
\alias{plot_pe}
\title{Plot Faith's Phylogenetic Endemism.}
\usage{
plot_pe(
  pb,
  palette = "viridis",
  border = "grey60",
  linewidth = 0.15,
  na_colour = "grey90",
  legend_title = "Phylogenetic Endemism"
)
}
\arguments{
\item{pb}{A \code{pb_project}.}

\item{palette}{Viridis colour palette.}

\item{border}{Polygon border colour.}

\item{linewidth}{Polygon border linewidth.}

\item{na_colour}{Colour used for missing values.}

\item{legend_title}{Optional legend title.}
}
\value{
A ggplot object.
}
\description{
Convenience wrapper around \code{\link[=plot_metric]{plot_metric()}} with \code{metric = "pe"}.
}
