% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_community.R
\name{prepare_community}
\alias{prepare_community}
\title{Prepare community data}
\usage{
prepare_community(pb)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}
}
\value{
Updated \code{pb_project}.
}
\description{
Builds internal lookup tables used by downstream biodiversity metrics.
}
