% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tree.R
\name{prepare_tree}
\alias{prepare_tree}
\title{Prepare a phylogenetic tree}
\usage{
prepare_tree(pb, verbose = TRUE, force = FALSE)
}
\arguments{
\item{pb}{A validated pb_project.}

\item{verbose}{Logical. Print progress messages?}

\item{force}{Logical. Continue even if tree validation failed.
Intended for development and validation workflows only.}
}
\value{
Updated pb_project.
}
\description{
Builds internal index structures used by downstream PhyloBasins algorithms.
}
