% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_community.R
\name{read_community}
\alias{read_community}
\title{Read community matrix}
\usage{
read_community(pb, file, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{file}{Path to the community matrix.}

\item{verbose}{Logical. Should progress messages be printed?}
}
\value{
Updated \code{pb_project}.
}
\description{
Imports a community (presence/absence or abundance) matrix and stores it in
a \code{pb_project}.
}
\details{
The first column must contain site names.
Remaining columns correspond to taxa.
}
