% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geometry.R
\name{read_geometry}
\alias{read_geometry}
\title{Read basin geometry}
\usage{
read_geometry(pb, file = NULL, data = NULL, quiet = TRUE, verbose = TRUE)
}
\arguments{
\item{pb}{A pb_project object.}

\item{file}{Path to a vector file supported by sf::st_read().}

\item{data}{An sf object.}

\item{quiet}{Suppress sf messages.}

\item{verbose}{Print progress messages.}
}
\value{
Updated pb_project.
}
\description{
Import spatial geometry into a PhyloBasins project.
}
