% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tree.R
\name{read_tree}
\alias{read_tree}
\title{Read a phylogenetic tree}
\usage{
read_tree(pb, file)
}
\arguments{
\item{pb}{A \code{pb_project} object.}

\item{file}{Path to the tree file.}
}
\value{
Updated \code{pb_project}.
}
\description{
Imports a phylogenetic tree from a Newick or NEXUS file and stores it
in a \code{pb_project} object.
}
\details{
Supported file extensions are:
\itemize{
\item .nwk
\item .newick
\item .tre
\item .tree
\item .nex
\item .nexus
}
}
\examples{
\dontrun{
pb <- pb_project()
pb <- read_tree(pb, "tree.nwk")
}

}
