% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the complete PhyloBasins pipeline}
\usage{
run_pipeline(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{overwrite}{Logical. Recompute already completed steps?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Executes all preparation and metric-computation steps required to analyse a
community dataset.
}
\details{
Already completed steps are skipped automatically unless
\code{overwrite = TRUE}.
}
