% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_match}
\alias{safe_match}
\title{Safe match}
\usage{
safe_match(x, table)
}
\arguments{
\item{x}{Values to match.}

\item{table}{Lookup table.}
}
\value{
Integer vector.
}
\description{
A wrapper around \code{match()} that always returns integer indices.
}
\keyword{internal}
