% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{timestamp}
\alias{timestamp}
\title{Current timestamp}
\usage{
timestamp()
}
\value{
POSIXct object.
}
\description{
Returns the current system time.
}
\details{
Exists mainly to centralise timestamp creation.
}
\keyword{internal}
