% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_community_data.R
\name{validate_community_data}
\alias{validate_community_data}
\title{Validate community data}
\usage{
validate_community_data(pb)
}
\arguments{
\item{pb}{A \code{pb_project}.}
}
\value{
Updated \code{pb_project}.
}
\description{
Performs structural validation of the community matrix.
}
