% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_geometry.R
\name{validate_geometry}
\alias{validate_geometry}
\title{Validate project geometry}
\usage{
validate_geometry(pb, require_polygons = TRUE, verbose = TRUE)
}
\arguments{
\item{pb}{A pb_project object.}

\item{require_polygons}{Require polygon geometries.}

\item{verbose}{Print validation summary.}
}
\value{
Updated pb_project.
}
\description{
Validate the geometry component of a PhyloBasins project.
}
