% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_branch_matrix.R
\name{validate_site_branch_matrix}
\alias{validate_site_branch_matrix}
\title{Validate a site-branch matrix}
\usage{
validate_site_branch_matrix(x)
}
\arguments{
\item{x}{A \code{pb_site_branch_matrix}.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate a site-branch matrix
}
\keyword{internal}
