% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_tree_data.R
\name{validate_tree_data}
\alias{validate_tree_data}
\title{Validate tree data}
\usage{
validate_tree_data(pb)
}
\arguments{
\item{pb}{A \code{pb_project} object with a loaded tree.}
}
\value{
Updated \code{pb_project}.
}
\description{
Performs a series of biological and structural checks on the tree stored
in a \code{pb_project}. The function updates the validation information
contained in \code{pb$tree$validation}.
}
\details{
The tree itself is not modified.
}
\examples{
\dontrun{
pb <- pb_project()
pb <- read_tree(pb, "tree.nwk")
pb <- validate_tree_data(pb)
}

}
