% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site_branch_matrix.R
\name{build_site_branch_matrix}
\alias{build_site_branch_matrix}
\title{Build the site by branch matrix}
\usage{
build_site_branch_matrix(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{overwrite}{Logical. Rebuild an existing matrix?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Builds the sparse site by branch incidence matrix used by all downstream
phylogenetic diversity calculations.
}
