% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_all_metrics.R
\name{compute_all_metrics}
\alias{compute_all_metrics}
\title{Compute all phylogenetic diversity metrics}
\usage{
compute_all_metrics(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{overwrite}{Logical. Recompute existing metrics?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes all currently implemented diversity metrics in the correct order.
}
\details{
The function automatically skips metrics that have already been computed,
unless \code{overwrite = TRUE}.
}
