% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_ranges.R
\name{compute_branch_ranges}
\alias{compute_branch_ranges}
\title{Compute branch geographic ranges}
\usage{
compute_branch_ranges(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A \code{pb_project} object.}

\item{overwrite}{Logical. Overwrite an existing result?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Calculates the geographic range of every branch from the site-by-branch
matrix.
}
