% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_branch_ranges.R
\name{compute_branch_ranges}
\alias{compute_branch_ranges}
\title{Compute branch ranges}
\usage{
compute_branch_ranges(pb)
}
\arguments{
\item{pb}{A prepared \code{pb_project}.}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes the number of sites occupied by each branch (the union of the
occupied sites of all descendant species).
}
\details{
The resulting range sizes are stored in
\code{pb$branches$table$branch_range}.
}
