% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pe.R
\name{compute_pe}
\alias{compute_pe}
\title{Compute Phylogenetic Endemism}
\usage{
compute_pe(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{overwrite}{Logical. Recompute existing values?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes phylogenetic endemism (PE) for all communities.
}
