% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pe.R
\name{compute_pe}
\alias{compute_pe}
\title{Compute Phylogenetic Endemism}
\usage{
compute_pe(pb, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{pb}{A \code{pb_project} object.}

\item{overwrite}{Logical. Overwrite an existing result?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
Updated \code{pb_project}.
}
\description{
Computes Faith's Phylogenetic Endemism (PE) for every site.
}
