% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_table.R
\name{metric_table}
\alias{metric_table}
\title{Extract a metric table}
\usage{
metric_table(pb, metric)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}

\item{metric}{Character string specifying the metric.}
}
\value{
A data frame containing the requested metric.
}
\description{
Returns a standardised table containing the values of a computed metric.
}
