% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{pb_community}
\alias{pb_community}
\title{Create a community object}
\usage{
pb_community(
  matrix = NULL,
  sites = character(),
  taxa = character(),
  file = NA_character_,
  loaded = FALSE,
  validation = list(valid = FALSE),
  metadata = list(),
  cache = list(taxa_index = NULL, site_index = NULL)
)
}
\arguments{
\item{matrix}{Community matrix.}

\item{sites}{Character vector of site names.}

\item{taxa}{Character vector of taxon names.}

\item{file}{Source file.}

\item{loaded}{Logical. Has the community matrix been loaded?}

\item{validation}{Validation information.}

\item{metadata}{Metadata associated with the community.}

\item{cache}{Internal cache.}
}
\value{
A \code{pb_community} object.
}
\description{
Creates a \code{pb_community} object.
}
