% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot a metric}
\usage{
plot_metric(
  pb,
  metric,
  palette = "viridis",
  border = "grey60",
  linewidth = 0.15,
  na_colour = "grey90",
  legend_title = NULL,
  ...
)
}
\arguments{
\item{pb}{A \code{pb_project} object.}

\item{metric}{Metric name.}

\item{palette}{Colour palette.}

\item{border}{Border colour.}

\item{linewidth}{Polygon border width.}

\item{na_colour}{Colour for missing values.}

\item{legend_title}{Legend title.}

\item{...}{Additional arguments passed to \code{\link[=plot_geometry]{plot_geometry()}}.}
}
\value{
A ggplot object.
}
\description{
Generic plotting function used by the metric wrappers.
}
