% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tree.R
\name{prepare_tree}
\alias{prepare_tree}
\title{Prepare a phylogenetic tree}
\usage{
prepare_tree(pb)
}
\arguments{
\item{pb}{A validated \code{pb_project}.}
}
\value{
Updated \code{pb_project}.
}
\description{
Builds internal index structures used by downstream PhyloBasins
algorithms. No biodiversity metrics are calculated.
}
