% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_community.R
\name{read_community}
\alias{read_community}
\title{Read a community matrix}
\usage{
read_community(pb, x, row_names = TRUE)
}
\arguments{
\item{pb}{A \code{pb_project}.}

\item{x}{Community matrix. May be
\itemize{
\item a matrix
\item a data.frame
\item a character string giving the path to a csv file.
}}

\item{row_names}{Should the first column be interpreted as site names?}
}
\value{
Updated \code{pb_project}.
}
\description{
Reads a site × species community matrix from a matrix,
data.frame or csv file.
}
